refactoring: move typemap_macro::tests into separate file
